requirements.txt
absl-py==0.11.0
autopep8==1.5.5
backcall==0.2.0
boto3==1.17.27
botocore==1.20.27
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
docker-pycreds==0.4.0
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.1
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.7.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.5
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
s3transfer==0.3.4
scikit-image==0.17.2
scipy==1.5.4
sentry-sdk==1.0.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminaltables==3.1.0
tifffile==2020.9.3
toml==0.10.2
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.59.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.10.21
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
